refactor: move the MainViewModel over to the pusher module
Move the ViewModel over ready for creating a fragment in that module
