+++ app/build.gradle
@@ -41,8 +41,6 @@ dependencies {
     implementation "androidx.appcompat:appcompat:1.3.0"
     implementation "androidx.constraintlayout:constraintlayout:2.0.4"
     implementation "androidx.core:core-ktx:1.5.0"
-    implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.3.1"
-    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.3.1"
     implementation "com.chesire.lintrules:lint-gradle:1.2.6"
     implementation "com.chesire.lintrules:lint-xml:1.2.6"
     implementation "com.google.android.material:material:1.3.0"
