+++ feature/pusher/src/main/java/com/chesire/pushie/pusher/PusherFragment.kt
@@ -1,0 +1,4 @@
+package com.chesire.pushie.pusher
+
+class PusherFragment {
+}
